Fixes steam installscript aborting
To enable/disable gallium nine, the appropriate registry key is either
added or deleted. This will print a harmless error about the key not
existing. However this might be interpreted by steam as a failure and
thus will abort dependency installation.

Any output is forwarded to /dev/null now.

Should fix #6 and #7